HS-926: address PR comments
