+++ metrics-processor/main/src/main/java/org/opennms/horizon/flows/FlowsApplicationConfig.java
@@ -29,6 +29,8 @@
 package org.opennms.horizon.flows;
 
 import com.codahale.metrics.MetricRegistry;
+import io.grpc.ManagedChannel;
+import io.grpc.ManagedChannelBuilder;
 import org.opennms.horizon.flows.cache.CacheConfig;
 import org.opennms.horizon.flows.classification.ClassificationEngine;
 import org.opennms.horizon.flows.classification.ClassificationRuleProvider;
@@ -75,6 +77,23 @@ public class FlowsApplicationConfig {
 
     @Value("${flows.nodeCache.recordStats:true}")
     private boolean nodeCacheRecordStats;
+
+    @Value("${grpc.url.inventory}")
+    private String inventoryGrpcAddress;
+    @Value("${grpc.server.deadline:60000}")
+    private long deadline;
+
+    @Bean
+    public ManagedChannel createInventoryChannel() {
+        return ManagedChannelBuilder.forTarget(inventoryGrpcAddress)
+            .keepAliveWithoutCalls(true)
+            .usePlaintext().build();
+    }
+
+    @Bean(destroyMethod = "shutdown", initMethod = "initialStubs")
+    public InventoryClient createInventoryClient(ManagedChannel channel) {
+        return new InventoryClient(channel, deadline);
+    }
 
     @Bean
     public CacheConfig createCacheConfig(final MetricRegistry metricRegistry) {
